Remove stuff from genesis_data, move ThemeStorage to its own file, use that everywhere
In general I think I should move the code to have a `data` folder, that has the storage handlers. Remove the logic from the FileBackend. And for API stuff that is re-used should be assigned a data handler or a helper, MVC-style, or should just be moved to the routes directly.

Also fix a number of small issues for dojo